Added parameter seach test

+++ hw/include/Utils.hpp
@@ -33,10 +33,10 @@ constexpr int ceil_deiv(float a, float b)
     return ceil(a / b);
 }
 
-constexpr int nearest_to_P(int n, int P) 
+constexpr int nearest_to_P(int n, int P)
 {
 
-    if (n % P == 0) 
+    if (n % P == 0)
     {
         return n;
     }
@@ -44,24 +44,24 @@ constexpr int nearest_to_P(int n, int P)
     return static_cast<int>(ceil(static_cast<double>(n) / P)) * P;
 }
 
-inline uint32_t set_bit(uint32_t aNumber, size_t anIndex) {
+inline uint64_t set_bit(uint64_t aNumber, size_t anIndex) {
     return aNumber | (1U << anIndex);
 }
 
-inline bool get_bit(uint32_t aNumber, size_t anIndex) {
+inline bool get_bit(uint64_t aNumber, size_t anIndex) {
     return (aNumber >> anIndex) & 1U;
 }
 
-static inline unsigned int initialize_rng() 
+static inline unsigned int initialize_rng()
 {
     unsigned int seed = static_cast<unsigned int>(std::chrono::system_clock::now().time_since_epoch().count());
     std::cout << "Random Seed for this run: " << seed << std::endl;  // Print the seed
-    return seed; 
+    return seed;
 }
 
 template <typename T>
 void compareQueues(std::queue<T>& aExpected, std::queue<T>& aActual, bool aCheckEquality = true) {
-    
+
     std::function<void(const T & aExpected, const T & aActual)> myWhileFunction = [](const T & aExpected, const T & aActual)
     {
         EXPECT_EQ(aExpected, aActual) << "Queue mismatch!";
@@ -76,17 +76,17 @@ void compareQueues(std::queue<T>& aExpected, std::queue<T>& aActual, bool aCheck
         };
     }
 
-    while (!aExpected.empty() && !aActual.empty()) 
+    while (!aExpected.empty() && !aActual.empty())
     {
         T myExpectedValue = aExpected.front();
         T myActualValue = aActual.front();
-        
+
         myWhileFunction(myExpectedValue, myActualValue);
 
         aExpected.pop();
         aActual.pop();
     }
-    
+
     EXPECT_EQ(aExpected.size(), 0);
     EXPECT_EQ(aActual.size(), 0);
     EXPECT_EQ(aActual.size(), 0);
